# vent/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: Vent Backend
host: 'localhost:5000'
basePath: /
...
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /users/register:
     post: 
       tags:
         - user
       summary: register a new user
       description: ''
       consumes:
        - application/json
       produces:
        - application/json
       parameters:
        - name: name
          in: body
        - name: email
          in: body
          required: true
        - name: username
          in: body
          required: true
        - name: password
          in: body
          required: true
       responses:
         default:
           description: successful operation
  /users/authenticate:
    post: 
      tags:
        - user
      summary: login authentication
      description: ''
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: username
          in: query
          type: string
          required: true
        - name: password
          in: query
          type: password
          required: true
      responses:
         default:
           description: successful operation
...
  /users/competSubscribe:
     post:
       tags:
         - user
       summary: subscribtion for a compettion
       description: ''
       consumes:
        - application/json
       produces:
        - application/json
       parameters:
        - name: userid
          in: query
          description: user id
          type: string
          required: true
        - name: competid
          in: query
          description: competition id
          type: string
          required: true
       responses:
         default:
           description: successful operation
  /users/competUnsubscribe:
     post:
       tags:
         - user
       summary: unsubscribe for a compettion
       description: ''
       consumes:
        - application/json
       produces:
        - application/json
       parameters:
        - name: userid
          in: query
          description: user id
          type: string
          required: true
        - name: competid
          in: query
          description: competition id
          type: string
          required: true
       responses:
         default:
           description: successful operation
  /users/teamSubscribe:
     post:
       tags:
         - user
       summary: subscribe for a team
       description: ''
       consumes:
        - application/json
       produces:
        - application/json
       parameters:
        - name: userid
          in: query
          description: user id
          type: string
          required: true
        - name: competid
          in: query
          description: team id
          type: string
          required: true
       responses:
         default:
           description: successful operation
  /users/teamUnsubscribe:
     post:
       tags:
         - user
       summary: unsubscribe for a team
       description: ''
       consumes:
        - application/json
       produces:
        - application/json
       parameters:
        - name: userid
          in: query
          description: user id
          type: string
          required: true
        - name: competid
          in: query
          description: team id
          type: string
          required: true
       responses:
         default:
           description: successful operation
  /users/resetpw:
     post:
       tags:
         - user
       summary: subscribe for a team
       description: ''
       consumes:
        - application/json
       produces:
        - application/json
       parameters:
        - name: username
          in: query
          description: username
          type: string
          required: true
       responses:
         default:
           description: successful operation
  /users/reset/{token}:
     post:
       tags:
         - user
       summary: reset pw by setting new password
       description: ''
       consumes:
        - application/json
       produces:
        - application/json
       parameters:
        - name: token
          in: path
          description: facebook token
          type: string
          required: true
        - name: competid
          in: query
          description: password
          type: string
          required: true
       responses:
         default:
           description: successful operation
...